Remove blueprint dashboard
