+++ builder/templates/common/_base.html
@@ -3,12 +3,15 @@
 <head>
     <meta charset="UTF-8">
     <title>PRODAM Builder :: {% block title %}{% endblock %}</title>
-    <link href="{{ url_for('static', filename='dashboard/style.css') }}" rel="stylesheet">
     <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.7/css/bootstrap.min.css", rel="stylesheet">
+    <link href="{{ url_for('static', filename='dashboard/style.css') }}" rel="stylesheet">
 </head>
 <div>
     {% include "common/_menu.html" %}
-    {% block layout %}{% endblock %}
+    <div class="row">
+        <div class="col-md-10 col-md-offset-1">{% include "common/_messages.html" %}</div>
+    </div>
+    <div class="row">{% block layout %}{% endblock %}</div>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.0/jquery.min.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.7/js/bootstrap.min.js"></script>
 </body>
