+++ builder/templates/common/_menu.html
@@ -7,12 +7,12 @@
       <span class="icon-bar"></span>
       <span class="icon-bar"></span>
     </button>
-    <a class="navbar-brand text-hide" href="{{ url_for('dashboard.home') }}">Brand Text
+    <a class="navbar-brand text-hide" href="{{ url_for('security.dashboard') }}">Brand Text
     </a>
   </div>
   <div id="navbar6" class="navbar-collapse collapse">
     <ul class="nav navbar-nav navbar-left">
-      <li><a href="{{ url_for('dashboard.home') }}"><span class="glyphicon glyphicon-home"></span> Home</a></li>
+      <li><a href="{{ url_for('security.dashboard') }}"><span class="glyphicon glyphicon-home"></span> Home</a></li>
       {% if current_user.superuser or current_user.admin %}
         <li><a href="{{ url_for('users.list_users') }}"><span class="glyphicon glyphicon-user"></span> Usuários</a></li>
       {% endif %}
@@ -22,7 +22,7 @@
       <li><a href="{{ url_for('core.list_projects') }}"><span class="glyphicon glyphicon-th-list"></span> Projetos</a></li>
     </ul>
     <ul class="nav navbar-nav navbar-right">
-      <li><a href="{{ url_for('dashboard.change_password') }}"><span class="glyphicon glyphicon-cog"></span> Alterar Senha</a></li>
+      <li><a href="{{ url_for('security.change_password') }}"><span class="glyphicon glyphicon-cog"></span> Alterar Senha</a></li>
       <li><a href="{{ url_for('security.logout') }}"><span class="glyphicon glyphicon-off"></span> Logout</a></li>
     </ul>
   </div>
