+++ builder/templates/security/change-password.html
@@ -8,7 +8,6 @@
     <hr>
     <form action="" id="resetPasswordForm" method="POST">
         {{ form.hidden_tag() }}
-        {% include "common/_messages.html" %}
         <div class="row">
             <div class="col-md-5 col-md-offset-1">
               <div class="form-group col-md-12">
@@ -30,7 +29,7 @@
         </div>
         <hr>
         <div class="text-center">
-          <a href="{{ url_for('dashboard.home') }}" class="btn btn-default">Voltar</a>
+          <a href="{{ url_for('security.dashboard') }}" class="btn btn-default">Voltar</a>
 
           <button type="submit" class="btn btn-success">
             <i class="glyphicon glyphicon-floppy-disk"></i> Salvar
