+++ builder/templates/users/add-user.html
@@ -24,7 +24,6 @@
 <form action="" id="userForm" method="POST">
   {{ form.hidden_tag() }}
 
-  {% include "common/_messages.html" %}
   <div class="form-group input-group col-md-12">
     <span class="input-group-addon"><i class="glyphicon glyphicon-user"></i></span>
     {{ form.username(placeholder=form.username.description, class='form-control', type='text', size='20') }}
