+++ builder/templates/users/details-user.html
@@ -3,8 +3,6 @@
 {% block title %}Detalhes do usuário{% endblock %}
 
 {% block content %}
-{% include "common/_messages.html" %}
-
 <h2 class="text-center">Detalhes do usuário</h2>
 
 <hr>
