+++ builder/utils.py
@@ -2,11 +2,9 @@
 def register_blueprints(app):
     """Shortcut to register all blueprints of system"""
     from builder.views.security import blueprint as security_blueprint
-    from builder.views.dashboard import blueprint as dashboard_blueprint
     from builder.views.users import blueprint as users_blueprint
     from builder.views.core import blueprint as core_blueprint
     app.register_blueprint(security_blueprint)
-    app.register_blueprint(dashboard_blueprint, url_prefix='/dashboard')
     app.register_blueprint(users_blueprint, url_prefix='/users')
     app.register_blueprint(core_blueprint, url_prefix='/core')
     return app
